requirements.txt
numpy
flask
scikit-learn
Pillow==9.5.0
tqdm
pandas
autolog
opencv
imutils
paddleocr>=2.0.1
paddlepaddle-gpu==2.5.0.post118 -f https://www.paddlepaddle.org.cn/whl/linux/mkl/avx/stable.html